requirements.txt
PyQt6>=6.4.0
qianfan>=0.3.0
PyYAML>=6.0
markdown>=3.4.0
requests>=2.28.0
